requirements.txt
pandas
numpy
scikit-learn
joblib
typer
pyyaml
